[Create and Update] created some dtos in auth service and modified the routes to use the headers in all services

+++ cart-service/src/main/java/com/backend/cart_service/controller/CartController.java
@@ -3,14 +3,12 @@ package com.backend.cart_service.controller;
 import com.backend.cart_service.utils.dto.CartDetailsResponse;
 import com.backend.cart_service.model.Cart;
 import com.backend.cart_service.service.CartService;
+import com.backend.cart_service.utils.exceptions.exps.IllegalAccessException;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 
 @Controller
 @AllArgsConstructor
@@ -20,9 +18,19 @@ public class CartController {
     private final CartService cartService;
 
     @GetMapping("/getCartDetails/cartId")
-    public ResponseEntity<CartDetailsResponse> getCartDetails(@RequestParam Long cartId) {
+    public ResponseEntity<CartDetailsResponse> getCartDetails(
+            @RequestHeader("X-Ref-Id") Long userId,
+            @RequestHeader("X-Role") String role,
+            @RequestParam Long cartId
+    ) {
 
-        CartDetailsResponse response = cartService.getCompleteCartDetails(cartId);
+        if (!role.equals("USER")) {
+            return ResponseEntity
+                    .status(HttpStatus.FORBIDDEN)
+                    .build();
+        }
+
+        CartDetailsResponse response = cartService.getCompleteCartDetails(cartId, userId);
         cartService.deleteCart(cartId);
 
         return ResponseEntity.ok(response);
