+++ restaurant-service/src/main/java/com/backend/restaurant_service/service/MenuItemService.java
@@ -6,9 +6,9 @@ import com.backend.restaurant_service.model.MenuItem;
 import com.backend.restaurant_service.model.Restaurant;
 import com.backend.restaurant_service.repository.MenuItemRepository;
 import com.backend.restaurant_service.repository.RestaurantRepository;
+import com.backend.restaurant_service.utils.exceptions.exps.IllegalRestaurantAccessException;
 import com.backend.restaurant_service.utils.exceptions.exps.MenuItemNotFoundException;
 import com.backend.restaurant_service.utils.exceptions.exps.RestaurantNotFoundException;
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -31,19 +31,23 @@ public class MenuItemService {
         menuItemRepository.save(menuItem);
     }
 
-    public void deleteMenuItem(Long menuItemId) {
+    public void deleteMenuItem(Long menuItemId, Long restaurantId) {
 
         MenuItem menuItem = menuItemRepository.findById(menuItemId)
                         .orElseThrow(() -> new MenuItemNotFoundException("Menu-item not found with id: " + menuItemId));
+
+        if(!menuItem.getRestaurant().getId().equals(restaurantId)) throw new IllegalRestaurantAccessException("You can modify this restaurant");
 
         menuItem.setUpdatedAt(LocalDateTime.now());
         menuItemRepository.deleteById(menuItemId);
     }
 
-    public void toggleAvailability(Long menuItemId) {
+    public void toggleAvailability(Long menuItemId, Long restaurantId) {
 
         MenuItem menuItem = menuItemRepository.findById(menuItemId)
                 .orElseThrow(() -> new MenuItemNotFoundException("Menu-item not found with id: " + menuItemId));
+
+        if(!menuItem.getRestaurant().getId().equals(restaurantId)) throw new IllegalRestaurantAccessException("You can modify this restaurant");
 
         menuItem.setIsAvailable(!menuItem.getIsAvailable());
         menuItem.setUpdatedAt(LocalDateTime.now());
